% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows.R
\name{split.grouped_df}
\alias{split.grouped_df}
\title{Split grouped_df}
\usage{
\method{split}{grouped_df}(x, ...)
}
\arguments{
\item{x}{A grouped_df}

\item{...}{Other arguments passed to \code{split}, if specified you must
supply a factor as with \code{base::split()}.`}
}
\value{
A list of data frames
}
\description{
Split a \code{grouped_df} according to its grouping variables.
}
\examples{
mtcars \%>\%
 dplyr::group_by(cyl) \%>\%
 split()
}
